git-svn-id: https://svn.osgeo.org/geotools/geotools/trunk@12987 e5c1c795-43da-0310-a71f-fac65c449510

+++ gt/plugin/arcgrid/src/org/geotools/gce/arcgrid/ArcGridRaster.java
@@ -391,7 +391,7 @@ public class ArcGridRaster {
                 }
 
                 // set the value at x,y in band 0 to the parsed value
-                raster.setSample(x, y, 0, d);
+                raster.setSample(x, y, 0,(float) d);
             }
         }
